Moved Class Color into 'model' directory
